+++ packages/buidler-truffle5/.gitignore
@@ -1,4 +1,16 @@
+# Node modules
+/node_modules
+
+# Compilation output
+/build-test/
 /dist
+
+# Code coverage artifacts
+/coverage
+/.nyc_output
+
+# Below is Github's node gitignore template, except for dotenv's entries as it's used by mocha to pass flags to ts-node,
+# and ignoring the node_modules part, as it'd ignore every node_modules, and we have some for testing
 
 # Logs
 logs
@@ -6,6 +18,10 @@ logs
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+lerna-debug.log*
+
+# Diagnostic reports (https://nodejs.org/api/report.html)
+report.[0-9]*.[0-9]*.[0-9]*.[0-9]*.json
 
 # Runtime data
 pids
@@ -22,7 +38,7 @@ coverage
 # nyc test coverage
 .nyc_output
 
-# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
+# Grunt intermediate storage (https://gruntjs.com/creating-plugins#storing-task-files)
 .grunt
 
 # Bower dependency directory (https://bower.io/)
@@ -35,7 +51,7 @@ bower_components
 build/Release
 
 # Dependency directories
-node_modules/
+#node_modules/
 jspm_packages/
 
 # TypeScript v1 declaration files
@@ -56,5 +72,28 @@ typings/
 # Yarn Integrity file
 .yarn-integrity
 
+# dotenv environment variables file
+#.env
+#.env.test
+
+# parcel-bundler cache (https://parceljs.org/)
+.cache
+
 # next.js build output
 .next
+
+# nuxt.js build output
+.nuxt
+
+# vuepress build output
+.vuepress/dist
+
+# Serverless directories
+.serverless/
+
+# FuseBox cache
+.fusebox/
+
+# DynamoDB Local files
+.dynamodb/
+
