+++ packages/buidler-truffle5/package.json
@@ -2,48 +2,49 @@
   "name": "@nomiclabs/buidler-truffle5",
   "version": "1.0.0-beta.1",
   "description": "Truffle 5 Buidler compatibility plugin",
-  "repository": "github:nomiclabs/buidler-truffle5",
+  "repository": "github:nomiclabs/buidler",
+  "homepage": "https://github.com/nomiclabs/buidler/packages/buidler-truffle5",
   "author": "Nomic Labs SRL",
   "license": "MIT",
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "keywords": [
+    "ethereum",
+    "smart-contracts",
+    "buidler",
+    "buidler-plugin",
+    "truffle",
+    "truffle-contract"
+  ],
   "scripts": {
-    "lint:fix": "prettier --write 'src/**/*.{js,ts}' 'test/**/*.{js,ts}' && tslint --fix --config tslint.json --project tsconfig.json",
-    "lint": "tslint --config tslint.json --project tsconfig.json",
-    "test": "scripts/run-tests.sh",
-    "build": "tsc",
-    "buidl": "tsc",
-    "watch": "tsc -w"
+    "lint:fix": "../../node_modules/.bin/prettier --write 'src/**/*.{js,ts}' 'test/**/*.{js,ts}' && npm run lint-src -- --fix && npm run lint-tests -- --fix",
+    "lint": "npm run lint-src && npm run lint-tests",
+    "lint-tests": "../../node_modules/.bin/tslint --config tslint.json --project ./tsconfig.json",
+    "lint-src": "../../node_modules/.bin/tslint --config tslint.json --project src/tsconfig.json",
+    "test": "./scripts/run-tests.sh",
+    "build": "../../node_modules/.bin/tsc --build src",
+    "build-test": "../../node_modules/.bin/tsc --build .",
+    "clean": "rm -rf dist build-test"
   },
   "files": [
-    "dist/src/",
+    "dist/",
     "src/",
     "LICENSE",
     "README.md"
   ],
   "dependencies": {
-    "@nomiclabs/buidler-web3": "^1.0.0-beta.1",
     "chai": "4.2.0",
     "truffle-contract": "4.0.4"
   },
   "devDependencies": {
+    "@nomiclabs/buidler-web3": "^1.0.0-beta.1",
+    "@nomiclabs/buidler": "^1.0.0-beta.2",
     "@types/chai": "^4.1.7",
-    "@types/fs-extra": "^5.0.4",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^8.10.38",
-    "dotenv": "^6.2.0",
-    "ganache-cli": "^6.2.5",
-    "mocha": "^5.1.1",
-    "prettier": "^1.15.3",
-    "source-map-support": "^0.5.9",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.12.0",
-    "tslint-config-prettier": "^1.17.0",
-    "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "^3.2.2"
+    "web3": "1.0.0-beta.37"
   },
   "peerDependencies": {
-    "@nomiclabs/buidler": "^1.0.0-beta.1",
+    "@nomiclabs/buidler-web3": "^1.0.0-beta.1",
+    "@nomiclabs/buidler": "^1.0.0-beta.2",
     "web3": "1.0.0-beta.37"
   }
 }
