+++ packages/buidler-truffle5/src/type-extensions.d.ts
@@ -1,0 +1,9 @@
+import "@nomiclabs/buidler/types";
+
+import { TruffleEnvironmentArtifacts } from "./artifacts";
+
+declare module "@nomiclabs/buidler/types" {
+  export interface BuidlerRuntimeEnvironment {
+    artifacts: TruffleEnvironmentArtifacts;
+  }
+}
